Added tests for base_product short_description

+++ src/models/base_product.rs
@@ -70,8 +70,8 @@ pub struct BaseProduct {
 }
 
 impl BaseProduct {
-    pub const MAX_LENGTH_SHORT_DESCRIPTION: usize = 170;
-    pub const MAX_LENGTH_LONG_DESCRIPTION: usize = 8000;
+    pub const MAX_LENGTH_SHORT_DESCRIPTION: u64 = 170;
+    pub const MAX_LENGTH_LONG_DESCRIPTION: u64 = 8000;
 }
 
 impl From<BaseProductRaw> for BaseProduct {
